pyproject.toml
[project]
name = "weather-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-bigquery>=1.11.1",
    "google-cloud-bigquery>=3.41.0",
    "google-cloud-storage>=3.1.1",
    "pandas>=3.0.2",
    "pyarrow>=23.0.1",
    "pyspark>=4.1.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
